bump(github.com/google/cadvisor): d937b68c57b7cd9187f77bfd30ee1629134ad337
This required some manual edits to
`k8s.io/kubernetes/pkg/kubelet/cm/util/cgroups_linux.go`, because some
function names changed in a newer version of
`github.com/opencontainers/runc` used by the newer version of cadvisor.

It also required some casting in
`k8s.io/kubernetes/pkg/kubelet/cm/cgroup_manager_linux.go`, since some
types in `github.com/opencontainers/runc` changed from `int64` to
`uint64`. It _should_ be fine, since we're converting `int64`s that
should probably never be negative to `uint64`s prior to assigning the
values to the runc struct.

+++ pkg/kubelet/cm/cgroup_manager_linux.go
@@ -332,16 +332,16 @@ func (m *cgroupManagerImpl) toResources(resourceConfig *ResourceConfig) *libcont
 		return resources
 	}
 	if resourceConfig.Memory != nil {
-		resources.Memory = *resourceConfig.Memory
+		resources.Memory = uint64(*resourceConfig.Memory)
 	}
 	if resourceConfig.CpuShares != nil {
-		resources.CpuShares = *resourceConfig.CpuShares
+		resources.CpuShares = uint64(*resourceConfig.CpuShares)
 	}
 	if resourceConfig.CpuQuota != nil {
 		resources.CpuQuota = *resourceConfig.CpuQuota
 	}
 	if resourceConfig.CpuPeriod != nil {
-		resources.CpuPeriod = *resourceConfig.CpuPeriod
+		resources.CpuPeriod = uint64(*resourceConfig.CpuPeriod)
 	}
 	return resources
 }
